setup/requirements.txt
cassandra-driver
ipython
ipython-extensions
